+++ lib/swaggerValidator/modelValidator.ts
@@ -762,6 +762,15 @@ export class SwaggerExampleValidator {
       headers["content-type"]?.split(";")[0] ||
       (isRequest ? undefined : "application/octet-stream");
     if (contentType !== undefined && !allowedContentTypes.includes(contentType)) {
+      // in some cases, produces value could have colon in type like 'application/json;odata=minimalmetadata'
+      for (const allowedContentType of allowedContentTypes) {
+        if (allowedContentType.includes(";")) {
+          const subAllowedContentType = allowedContentType.split(";")[0];
+          if (subAllowedContentType.includes(contentType)) {
+            return;
+          }
+        }
+      }
       this.errors.push(
         this.issueFromErrorCode(
           operationId,