+++ ChangeLog.md
@@ -6,7 +6,7 @@
 - Extends format types in AJV schema validation
 - Bug fix - check discriminator property type must be string
 - Upgrade autorest scheams about x-ms-enum change
-
+- Model/Live validation - fix content-type validation and unknown format issue during validator building
 
 ## 10/26/2021 2.10.1